--- app/src/App.jsx
@@ -1,7 +1,0 @@
-import './App.css'
-
-function App() {
-  return <div>'Hello World'</div>
-}
-
-export default App
