+++ src/actions/init.js
@@ -1,0 +1,5 @@
+import { options } from './options.js'
+
+export async function initActions(params) {
+  params.map((name) => options[name].action('none'))
+}
