+++ src/actions/options.js
@@ -1,0 +1,7 @@
+import initializeConfig from '../purposes/linting/bin/initialize.js'
+
+export const options = {
+  linting: {
+    action: initializeConfig
+  }
+}
