+++ src/purposes/linting/bin/initialize.js
