+++ src/purposes/linting/config-file.js
