+++ src/purposes/linting/customConfig.js
