+++ src/purposes/linting/getLocalEslintVersion.js
