+++ src/purposes/linting/getStyleGuideName.js
