+++ src/purposes/linting/hasEslintVersionConflict.js
