+++ src/purposes/linting/initializeConfig.js
