+++ src/purposes/linting/shared/logging.js
