+++ src/purposes/linting/writeFile.js
