+++ src/purposes/purposes.js
@@ -1,0 +1,24 @@
+import { blue, green, magenta, red } from 'kolorist'
+
+export const purposes = [
+  {
+    name: 'linting',
+    display: 'Linting',
+    color: blue
+  },
+  {
+    name: 'styling',
+    display: 'Styling',
+    color: red
+  },
+  {
+    name: 'testing',
+    display: 'Testing',
+    color: green
+  },
+  {
+    name: 'architecture',
+    display: 'Architecture',
+    color: magenta
+  }
+]
